Add cache for graphql-config loading
